requirements.txt
blobfile==3.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
editdistance==0.8.1
fairseq2==0.4.0
fairseq2n==0.4.0
ffmpeg==1.4
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.2.1
portalocker==3.1.1
psutil==5.9.8
pycparser==2.22
pycryptodomex==3.23.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sacrebleu==2.5.1
six==1.17.0
sonar-space==0.4.0
sounddevice==0.5.2
soundfile==0.13.1
sox==1.5.0
sympy==1.13.1
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.3.0
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
torcheval==0.0.7
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zipp==3.21.0
